+++ feature/notes/impl/src/main/kotlin/com/oussamateyib/thoth/feature/notes/impl/list/NoteListEvent.kt
@@ -6,7 +6,7 @@ import com.oussamateyib.thoth.core.model.data.NoteColor
 sealed class NoteListEvent {
     data class Order(val noteOrder: NoteOrder) : NoteListEvent()
     object ToggleSortSheet : NoteListEvent()
-    data class SelectNote(val id: Int) : NoteListEvent()
+    data class SelectNote(val id: Long) : NoteListEvent()
     object ClearSelection : NoteListEvent()
     object DeleteSelectedNotes : NoteListEvent()
     object RestoreDeletedNotes : NoteListEvent()
